requirements.txt
pandas
py-cord
python-dotenv
gspread
google-auth